fastapi-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
redis==5.0.8
httpx==0.27.2
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.12
boto3==1.35.0
openpyxl==3.1.5
pytz==2024.1
python-whois==0.8.0
feedparser==6.0.11
pyfcm==1.5.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2


